Consolidate bail detail classification
